+++ packages/generator-suit/CHANGELOG.md
@@ -1,5 +1,7 @@
 === HEAD
 
+* Switch to using Component(1) for development.
+
 === 0.1.1 (October 19, 2013)
 
 * Fix the generated README's description placeholder.